listener/requirements.txt
bleak>=0.22.0
paho-mqtt>=2.0.0
PyYAML>=6.0
